requirements.txt
aiogram==3.13.1
python-dotenv==1.0.0
aiofiles==23.2.1
gspread==5.12.0
google-auth==2.25.2

